Avoid wrapping a RuntimeException from wrapped main in an UndeclaredThrowableException
Also: Add an install target on the top-most Makefile, that installs modulec
and javahms as symlinks.  Symlink resolution is fixed to allow javahms
to find back to the hybridmodule home directory and find the no.ion.jhms
JHMS implementation.

modulec's -f now supports ending the jar path with a /, which means it will
create that directory if it does not exist.